+++ lib/Controller/WorkspaceController.php
@@ -25,75 +25,57 @@
 
 namespace OCA\Workspace\Controller;
 
-use OCA\Workspace\Service\Workspace\WorkspaceCheckService;
-use OCP\ILogger;
-use OCP\IRequest;
-use OCP\IUserManager;
-use OCP\IGroupManager;
-use OCP\AppFramework\Controller;
-use OCP\AppFramework\Http;
-use OCP\AppFramework\Http\JSONResponse;
-use OCA\Workspace\Db\Space;
-use OCA\Workspace\Db\SpaceMapper;
 use OCA\Workspace\AppInfo\Application;
 use OCA\Workspace\BadRequestException;
 use OCA\Workspace\CreateGroupException;
-use OCA\Workspace\Service\UserService;
-use OCA\Workspace\Service\SpaceService;
-use OCA\Workspace\Service\WorkspaceService;
 use OCA\Workspace\CreateWorkspaceException;
+use OCA\Workspace\Db\Space;
+use OCA\Workspace\Db\SpaceMapper;
+use OCA\Workspace\Service\SpaceService;
+use OCA\Workspace\Service\UserService;
+use OCA\Workspace\Service\Workspace\WorkspaceCheckService;
+use OCA\Workspace\Service\WorkspaceService;
+use OCP\AppFramework\Controller;
+use OCP\AppFramework\Http;
+use OCP\AppFramework\Http\JSONResponse;
+use OCP\IGroupManager;
+use OCP\ILogger;
+use OCP\IRequest;
+use OCP\IUserManager;
 
 class WorkspaceController extends Controller {
 
-    /** @var IGroupManager */
-    private $groupManager;
-
-    /** @var ILogger */
-    private $logger;
-
-    /** @var IUserManager */
-    private $userManager;
-
-    /** @var UserService */
-    private $userService;
-
-    /** @var SpaceService */
-    private $spaceService;
-
-    /** @var SpaceMapper */
-    private $spaceMapper;
-
-    /** @var WorkspaceService */
-    private $workspaceService;
-
-    /** @var WorkspaceCheckService */
-    private $workspaceCheck;
+    private IGroupManager $groupManager;
+    private ILogger $logger;
+    private IUserManager $userManager;
+    private SpaceMapper $spaceMapper;
+    private SpaceService $spaceService;
+    private UserService $userService;
+    private WorkspaceCheckService $workspaceCheck;
+    private WorkspaceService $workspaceService;
 
     public function __construct(
 		$AppName,
 		IGroupManager $groupManager,
 		ILogger $logger,
 		IRequest $request,
-		UserService $userService,
 		IUserManager $userManager,
 		SpaceMapper $mapper,
 		SpaceService $spaceService,
-		WorkspaceService $workspaceService,
-		WorkspaceCheckService $workspaceCheck
+		UserService $userService,
+		WorkspaceCheckService $workspaceCheck,
+		WorkspaceService $workspaceService
     )
     {
 	parent::__construct($AppName, $request);
-
 		$this->groupManager = $groupManager;
 		$this->logger = $logger;
-
-		$this->userManager = $userManager;
-		$this->userService = $userService;
-
 		$this->spaceMapper = $mapper;
 		$this->spaceService = $spaceService;
-		$this->workspaceService = $workspaceService;
+		$this->userManager = $userManager;
+		$this->userService = $userService;
 		$this->workspaceCheck = $workspaceCheck;
+		$this->workspaceService = $workspaceService;
     }
 
     /**
@@ -168,111 +150,6 @@ class WorkspaceController extends Controller {
                     'displayName' => $newSpaceUsersGroup->getDisplayName(),
                 ]
             ],
-            'statuscode' => Http::STATUS_CREATED,
-        ]);
-    }
-
-    /**
-     * @NoAdminRequired
-     * @GeneralManagerRequired
-     * @NoCSRFRequired
-     */
-    public function convertGroupfolderToSpace(string $spaceName, $groupfolder) {
-
-        if ( $spaceName === false ||
-            $spaceName === null ||
-            $spaceName === ''
-        ) {
-            throw new BadRequestException('spaceName must be provided');
-        }
-
-        $this->workspaceCheck->containSpecialChar($spaceName);
-
-        $spaceName = $this->deleteBlankSpaceName($spaceName);
-
-        if (gettype($groupfolder) === 'string') {
-			$groupfolder = json_decode($groupfolder, true);
-		}
-
-        $this->workspaceCheck->isExist($spaceName);
-
-        // #1 create the space
-        $space = new Space();
-        $space->setSpaceName($spaceName);
-        $space->setGroupfolderId($groupfolder['id']);
-        $space->setColorCode('#' . substr(md5(mt_rand()), 0, 6)); // mt_rand() (MT - Mersenne Twister) is taller efficient than rand() function.
-        $this->spaceMapper->insert($space);
-
-        if (is_null($space)) {
-            return new JSONResponse([
-                'statuscode' => Http::STATUS_BAD_REQUEST,
-                'message' => 'Error to create a space.',
-            ]);
-        }
-
-        // #2 create groups
-        $newSpaceManagerGroup = $this->groupManager->createGroup(Application::GID_SPACE . Application::ESPACE_MANAGER_01 . $space->getId());
-
-        if (is_null($newSpaceManagerGroup)) {
-            return new JSONResponse([
-                'statuscode' => Http::STATUS_BAD_REQUEST,
-                'message' => 'Error to create a Space Manager group.',
-            ]);
-        }
-
-        $newSpaceUsersGroup = $this->groupManager->createGroup(Application::GID_SPACE . Application::ESPACE_USERS_01 . $space->getId());
-
-        if (is_null($newSpaceUsersGroup)) {
-            return new JSONResponse([
-                'statuscode' => Http::STATUS_BAD_REQUEST,
-                'message' => 'Error to create a Space Users group.',
-            ]);
-        }
-
-        $newSpaceManagerGroup->setDisplayName(Application::ESPACE_MANAGER_01 . $space->getId());
-        $newSpaceUsersGroup->setDisplayName(Application::ESPACE_USERS_01 . $space->getId());
-
-        $groupsName = array_keys($groupfolder['groups']);
-
-        $groupsOfGroupfolder = [];
-        $users = [];
-        // To cheat the fomatUser method.
-        $groupfolder['groups'][$newSpaceUsersGroup->getGID()] = 31;
-        foreach($groupsName as $groupName) {
-            $group = $this->groupManager->get($groupName);
-
-            $usersOfAGroup = $group->getUsers();
-            foreach($usersOfAGroup as $user) {
-                $newSpaceUsersGroup->addUser($user);
-                $users[$user->getUID()] = $this->userService->formatUser($user, $groupfolder, 'user');
-            }
-            $groupsOfGroupfolder[$group->getGID()] = [
-                    'gid' => $group->getGID(),
-                    'displayName' => $group->getDisplayName(),
-                ];
-        }
-        $groups = array_merge(
-            $groupsOfGroupfolder,
-            [
-                $newSpaceManagerGroup->getGID() => [
-                    'gid' => $newSpaceManagerGroup->getGID(),
-                    'displayName' => $newSpaceManagerGroup->getDisplayName(),
-                ],
-                $newSpaceUsersGroup->getGID() => [
-                    'gid' => $newSpaceUsersGroup->getGID(),
-                    'displayName' => $newSpaceUsersGroup->getDisplayName(),
-                ]
-            ]
-        );
-
-		// #3 Returns result
-        return new JSONResponse ([
-            'space_name' => $space->getSpaceName(),
-            'id_space' => $space->getId(),
-            'folder_id' => $space->getGroupfolderId(),
-            'color' => $space->getColorCode(),
-            'groups' => $groups,
-            'users' => (object)$users,
             'statuscode' => Http::STATUS_CREATED,
         ]);
     }