+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refactoring the creation workspace part ( [#609](https://github.com/arawa/workspace/pull/609) )
+- Refactorised the converting groupfolders to workspace ( [#633](https://github.com/arawa/workspace/pull/633) )
 
 ### Style